refactor(tests): move reviewer-fixes tests to module-specific files
Redistribute 5 test classes from test_reviewer_fixes.py to files that
follow the test_{module_name}.py convention:

- TestLoadVcMemberNamePattern + TestGenerateVcMemberName
  → new test_virtual_chassis.py (tests import_utils/virtual_chassis.py)
- TestRenderDeviceSelectionEscape
  → test_view_wiring.py (rendering/escape quality checks)
- TestImportSingleDeviceLazyValidation
  → test_import_utils.py (tests import_utils/device_operations.py)
- TestCreatePlatformFullClean
  → test_sync_devices.py (tests views/sync/device_fields.py)

Delete test_reviewer_fixes.py (named after a process artifact, not a module).
